Organised components by their categories
